Move the JWT helpers out of guards and into auth/tokens
jwtreq.ts lived in auth/guards/ and held no guard. What it holds is the
cookie names, the token type enum, the payload schema, the four readers,
and sessionFromRequest, and the guards import from it rather than the
other way round. Opening auth/guards/ to find where a token comes off a
request only worked if you already knew jwtreq was the answer, which is
the thing the previous commit set out to fix.

It is auth/tokens.ts now, named for what is in it. The code is untouched;
28 files repoint their import and the spec moves with the file.
