+++ server/src/auth/auth.controller.ts
@@ -37,15 +37,15 @@ import { SignUpDto } from "./dto/sign-up.dto";
 import { SignInDto, SignInResponseDto, type TokenMode } from "./dto/signin.dto";
 import { AdminGuard } from "./guards/admin.guard";
 import { AuthGuard } from "./guards/auth.guard";
+import { RefreshTokenGuard } from "./guards/refresh.guard";
+import { Public } from "./public.decorator";
+import { SIGNUP_THROTTLE } from "./signup-throttle.config";
 import {
   extractGuestTokenFromCookie,
   extractRefreshTokenFromCookie,
   type JwtRequest,
   sessionFromRequest,
-} from "./guards/jwtreq";
-import { RefreshTokenGuard } from "./guards/refresh.guard";
-import { Public } from "./public.decorator";
-import { SIGNUP_THROTTLE } from "./signup-throttle.config";
+} from "./tokens";
 
 class TokenModeQuery {
   @ApiPropertyOptional({ enum: ["cookie", "header"] })
