+++ server/src/auth/user.decorator.ts
@@ -1,5 +1,5 @@
 import { createParamDecorator, ExecutionContext } from "@nestjs/common";
-import type { JwtPayload, JwtRequest } from "./guards/jwtreq";
+import type { JwtPayload, JwtRequest } from "./tokens";
 
 export const ReqUser = createParamDecorator(
   (_data: unknown, ctx: ExecutionContext): JwtPayload => {
