+++ server/src/contract/contract.controller.ts
@@ -13,8 +13,8 @@ import {
 import { ApiOkResponse } from "@nestjs/swagger";
 import { AdminGuard } from "src/auth/guards/admin.guard";
 import { AuthGuard } from "src/auth/guards/auth.guard";
-import type { JwtRequest } from "src/auth/guards/jwtreq";
 import { Public } from "src/auth/public.decorator";
+import type { JwtRequest } from "src/auth/tokens";
 import { PosthogService } from "src/posthog/posthog.service";
 import { ContractService } from "./contract.service";
 import {
