+++ server/src/messaging/conversation.controller.ts
@@ -14,7 +14,7 @@ import {
 import { ApiOkResponse, ApiTags } from "@nestjs/swagger";
 import { AdminGuard } from "src/auth/guards/admin.guard";
 import { AuthGuard } from "src/auth/guards/auth.guard";
-import type { JwtRequest } from "src/auth/guards/jwtreq";
+import type { JwtRequest } from "src/auth/tokens";
 import { PosthogService } from "src/posthog/posthog.service";
 import { ConversationService } from "./conversation.service";
 import {
