+++ server/src/push/push.controller.ts
@@ -2,7 +2,7 @@ import { Body, Controller, Post, Request, UseGuards } from "@nestjs/common";
 import { ApiOkResponse } from "@nestjs/swagger";
 import { InjectRepository } from "@nestjs/typeorm";
 import { AuthGuard } from "src/auth/guards/auth.guard";
-import type { JwtRequest } from "src/auth/guards/jwtreq";
+import type { JwtRequest } from "src/auth/tokens";
 import { Repository } from "typeorm";
 import { PushOpenedDto } from "./dto/push-opened.dto";
 import { Push } from "./push.entity";
