+++ server/src/share-urls/share-urls.controller.ts
@@ -15,7 +15,7 @@ import {
 import { ApiOkResponse } from "@nestjs/swagger";
 import { AdminGuard } from "src/auth/guards/admin.guard";
 import { AuthGuard } from "src/auth/guards/auth.guard";
-import type { JwtRequest } from "src/auth/guards/jwtreq";
+import type { JwtRequest } from "src/auth/tokens";
 import {
   CreateDuplicateShareLinkDto,
   CreateInviteDuplicateDto,
